Fix Test directory compiler warnings.

+++ src/test/java/tfw/immutable/ila/doubleila/DoubleIlaFromCastIntIlaTest.java
@@ -24,7 +24,7 @@ class DoubleIlaFromCastIntIlaTest {
         IntIla ila = IntIlaFromArray.create(array);
         DoubleIla targetIla = DoubleIlaFromArray.create(target);
         DoubleIla actualIla = DoubleIlaFromCastIntIla.create(ila);
-        final double epsilon = (double) 0.0;
+        final double epsilon = 0.0;
         DoubleIlaCheck.checkAll(
                 targetIla,
                 actualIla,
