+++ src/test/java/tfw/immutable/ila/intila/IntIlaFromCastCharIlaTest.java
@@ -24,7 +24,7 @@ class IntIlaFromCastCharIlaTest {
         CharIla ila = CharIlaFromArray.create(array);
         IntIla targetIla = IntIlaFromArray.create(target);
         IntIla actualIla = IntIlaFromCastCharIla.create(ila);
-        final int epsilon = (int) 0.0;
+        final int epsilon = 0;
         IntIlaCheck.checkAll(
                 targetIla,
                 actualIla,
