+++ src/test/java/tfw/immutable/ila/longila/LongIlaFromCastShortIlaTest.java
@@ -24,7 +24,7 @@ class LongIlaFromCastShortIlaTest {
         ShortIla ila = ShortIlaFromArray.create(array);
         LongIla targetIla = LongIlaFromArray.create(target);
         LongIla actualIla = LongIlaFromCastShortIla.create(ila);
-        final long epsilon = (long) 0.0;
+        final long epsilon = 0L;
         LongIlaCheck.checkAll(
                 targetIla,
                 actualIla,
